collapsible section styled and show icomponentprops first
