+++ apps/fabric-website-resources/src/components/pages/CalloutPage.tsx
@@ -3,4 +3,6 @@ import * as React from 'react';
 import { DemoPage } from '../DemoPage';
 import { CalloutPageProps } from 'office-ui-fabric-react/lib/components/Callout/Callout.doc';
 
-export const CalloutPage = (props: { isHeaderVisible: boolean }) => <DemoPage {...{ ...CalloutPageProps, ...props }} />;
+export const CalloutPage = (props: { isHeaderVisible: boolean }) => (
+  <DemoPage jsonDocs={require('../../../../../common/pages/Callout.page.json')} {...{ ...CalloutPageProps, ...props }} />
+);
