+++ apps/fabric-website-resources/src/components/pages/ComboBoxPage.tsx
@@ -3,5 +3,5 @@ import { DemoPage } from '../DemoPage';
 import { ComboBoxPageProps } from 'office-ui-fabric-react/lib/components/ComboBox/ComboBox.doc';
 
 export const ComboBoxPage = (props: { isHeaderVisible: boolean }) => (
-  <DemoPage {...{ ...ComboBoxPageProps, ...props }} />
+  <DemoPage jsonDocs={require('../../../../../common/pages/ComboBox.page.json')} {...{ ...ComboBoxPageProps, ...props }} />
 );
