+++ common/pages/Calendar.page.json
@@ -1,0 +1,3 @@
+{
+  "tables": []
+}
