--- common/pages/IScheme.page.json
@@ -1,72 +1,0 @@
-{
-  "tables": [
-    {
-      "kind": "interface",
-      "name": "IScheme",
-      "extendsTokens": [],
-      "descriptionHtml": "\n",
-      "members": [
-        {
-          "name": "disableGlobalClassNames",
-          "typeTokens": [
-            {
-              "text": "boolean"
-            }
-          ],
-          "descriptionHtml": "This setting is for a very narrow use case and you probably don't need to worry about, unless you share a environment with others that also use fabric. It is used for disabling global styles on fabric components. This will prevent global overrides that might have been set by other fabric users from applying to your components. When you set this setting to `true` on your theme the components in the subtree of your Customizer will not get the global styles applied to them.\n",
-          "deprecated": false
-        },
-        {
-          "name": "effects",
-          "typeTokens": [
-            {
-              "text": "IEffects"
-            }
-          ],
-          "descriptionHtml": "",
-          "deprecated": false
-        },
-        {
-          "name": "fonts",
-          "typeTokens": [
-            {
-              "text": "IFontStyles"
-            }
-          ],
-          "descriptionHtml": "",
-          "deprecated": false
-        },
-        {
-          "name": "isInverted",
-          "typeTokens": [
-            {
-              "text": "boolean"
-            }
-          ],
-          "descriptionHtml": "",
-          "deprecated": false
-        },
-        {
-          "name": "palette",
-          "typeTokens": [
-            {
-              "text": "IPalette"
-            }
-          ],
-          "descriptionHtml": "",
-          "deprecated": false
-        },
-        {
-          "name": "semanticColors",
-          "typeTokens": [
-            {
-              "text": "ISemanticColors"
-            }
-          ],
-          "descriptionHtml": "",
-          "deprecated": false
-        }
-      ]
-    }
-  ]
-}
