--- common/pages/ITheme.page.json
@@ -1,18 +1,0 @@
-{
-  "tables": [
-    {
-      "kind": "interface",
-      "name": "ITheme",
-      "extendsTokens": [
-        {
-          "text": "IScheme"
-        },
-        {
-          "text": " "
-        }
-      ],
-      "descriptionHtml": "\n",
-      "members": []
-    }
-  ]
-}
