+++ packages/experiments/src/components/CollapsibleSection/examples/CollapsibleSection.Basic.Example.tsx
@@ -7,7 +7,7 @@ export class CollapsibleSectionBasicExample extends React.Component<{}, {}> {
     return (
       <div>
         <FocusZone>
-          <CollapsibleSection key={1} defaultCollapsed={true} title="Title 1">
+          <CollapsibleSection key={1} defaultCollapsed={true} title={{ text: 'hi', style: null }}>
             Content 1
           </CollapsibleSection>
         </FocusZone>
