+++ packages/office-ui-fabric-react/src/components/Checkbox/Checkbox.types.ts
@@ -6,6 +6,7 @@ import { IKeytipProps } from '../../Keytip';
 
 /**
  * Checkbox class interface.
+ * {@docCategory Checkbox}
  */
 export interface ICheckbox {
   /** Gets the current checked state. */
@@ -17,6 +18,7 @@ export interface ICheckbox {
 
 /**
  * Checkbox properties.
+ * {@docCategory Checkbox}
  */
 export interface ICheckboxProps extends React.ButtonHTMLAttributes<HTMLElement | HTMLInputElement> {
   /**
@@ -122,6 +124,9 @@ export interface ICheckboxProps extends React.ButtonHTMLAttributes<HTMLElement |
   keytipProps?: IKeytipProps;
 }
 
+/**
+ * {@docCategory Checkbox}
+ */
 export interface ICheckboxStyleProps {
   theme: ITheme;
   className?: string;
@@ -131,6 +136,9 @@ export interface ICheckboxStyleProps {
   isUsingCustomLabelRender: boolean;
 }
 
+/**
+ * {@docCategory Checkbox}
+ */
 export interface ICheckboxStyles {
   /**
    * Style for the root element (a button) of the checkbox component in the default enabled/unchecked state.
